Fix API exceptions handler
